% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_functions.R
\name{get_and_store}
\alias{get_and_store}
\title{Retrieve tweets and store locally}
\usage{
get_and_store(query = "#rspatial", n_tweets = 20,
  tweets_file = "tweets_rspatial.rds",
  complete_tweets_file = "complete_tweets_rspatial.rds", log = TRUE,
  logfile = "rtweet_console.log")
}
\arguments{
\item{query}{search tweet query}

\item{n_tweets}{n_tweets}

\item{tweets_file}{tweets_file}

\item{complete_tweets_file}{complete_tweets_file}

\item{log}{log}

\item{logfile}{logfile}
}
\description{
Retrieve tweets and store locally
}
