% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tweets.R
\name{top_tweets}
\alias{top_tweets}
\title{Create a tweet for top retweets and other stats}
\usage{
top_tweets(
  all_tweets,
  save_dir = tempdir(),
  post_tweet = TRUE,
  top_number = 5,
  hashtag = "rspatial",
  fill = c("#1e73be", "#BF223C", "#79C698"),
  token = NULL
)
}
\arguments{
\item{all_tweets}{Table of tweets to explore}

\item{save_dir}{Path where to save images that will be tweet}

\item{post_tweet}{Logical. Whether to really tweet the content}

\item{top_number}{Numeric. Number of best retweeted tweets to show on the graph}

\item{hashtag}{Character. Which Twitter hashtag is followed}

\item{fill}{Vector of colors to be used in the graphics.
Length 1 for all the same, or length 3 resp. for:
top retweeted, tweets per day, contributors per day}

\item{token}{Every user should have their own Oauth (Twitter API) token.
By default token = NULL this function looks for the path to a saved Twitter
token via environment variables (which is what 'create_token()'
sets up by default during initial token creation).
For instruction on how to create a Twitter token see the tokens vignette,
i.e., 'vignettes("auth", "rtweet")'.}
}
\description{
Create a tweet for top retweets and other stats
}
