% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_last_month.R
\name{filter_month}
\alias{filter_month}
\title{Filter dataset for last month tweets only}
\usage{
filter_month(x, the_month, the_year)
}
\arguments{
\item{x}{Dataset of tweets}

\item{the_month}{Number of the month. Default to last month.}

\item{the_year}{Year to filter. Default to current year.}
}
\description{
Filter dataset for last month tweets only
}
