% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_info.R
\name{get_account_info}
\alias{get_account_info}
\title{Retrieve number of followers and tweets}
\usage{
get_account_info(user = "talk_rspatial", dir = ".",
  timeline_file = "timeline_rspatial.rds", log = TRUE,
  logfile = "rtweet_info.log")
}
\arguments{
\item{user}{user names or user IDs}

\item{dir}{Directory where to save data}

\item{timeline_file}{timeline filename relative to dir}

\item{log}{Logical. log file}

\item{logfile}{lofile name relative to dir}
}
\description{
Retrieve number of followers and tweets
}
