% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_functions.R
\name{get_and_store}
\alias{get_and_store}
\title{Retrieve tweets and store locally}
\usage{
get_and_store(
  query = "#rspatial",
  n_tweets = 20,
  dir = ".",
  tweets_file = "tweets_rspatial.rds",
  complete_tweets_file = "complete_tweets_rspatial.rds",
  log = TRUE,
  logfile = "rtweet_console.log",
  loop_pid_file = "loop_pid.log",
  token = NULL
)
}
\arguments{
\item{query}{search tweet query}

\item{n_tweets}{n_tweets}

\item{dir}{Path where everything will be saved}

\item{tweets_file}{tweets_file path name relative to dir}

\item{complete_tweets_file}{complete_tweets_file path name relative to dir}

\item{log}{Logical}

\item{logfile}{logfile path name relative to dir}

\item{loop_pid_file}{loop_pid_file path name relative to dir}

\item{token}{Every user should have their own Oauth (Twitter API) token.
By default token = NULL this function looks for the path to a saved Twitter
token via environment variables (which is what 'create_token()'
sets up by default during initial token creation).
For instruction on how to create a Twitter token see the tokens vignette,
i.e., 'vignettes("auth", "rtweet")'.}
}
\description{
Retrieve tweets and store locally
}
