% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_functions.R
\name{retweet_and_update}
\alias{retweet_and_update}
\title{Retweet and update the database of tweets to tweet}
\usage{
retweet_and_update(tweets_file = "tweets_rspatial.rds",
  complete_tweets_file = "complete_tweets_rspatial.rds", log = TRUE,
  logfile = "rtweet_console.log", loop_pid_file = "loop_pid.log",
  tweets_failed_file = "tweets_failed_rspatial.rds", n_tweets = 20,
  n_limit = 3, sys_sleep = 600, debug = FALSE)
}
\arguments{
\item{tweets_file}{tweets_file}

\item{complete_tweets_file}{complete_tweets_file}

\item{log}{Logical}

\item{logfile}{logfile}

\item{loop_pid_file}{loop_pid_file}

\item{tweets_failed_file}{tweets_failed_file}

\item{n_tweets}{n_tweets}

\item{n_limit}{n_limit}

\item{sys_sleep}{sys_sleep}

\item{debug}{Logical. Use TRUE to avoid tweeting while debugging}
}
\description{
Retweet and update the database of tweets to tweet
}
