% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tweets.R
\name{top_tweets}
\alias{top_tweets}
\title{Create a tweet for top retweets}
\usage{
top_tweets(all_tweets, save_dir = tempdir(), post_tweet = TRUE)
}
\arguments{
\item{all_tweets}{Table of tweets to explore}

\item{save_dir}{Path where to save images that will be tweet}

\item{post_tweet}{Logical. Whether to really tweet the content}
}
\description{
Create a tweet for top retweets
}
