% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tweets.R
\name{top_tweets}
\alias{top_tweets}
\title{Create a tweet for top retweets and other stats}
\usage{
top_tweets(
  all_tweets,
  save_dir = tempdir(),
  post_tweet = TRUE,
  top_number = 5,
  hashtag = "rspatial",
  fill = c("#1e73be", "#BF223C", "#79C698"),
  token = NULL,
  the_month_name
)
}
\arguments{
\item{all_tweets}{Table of tweets to explore}

\item{save_dir}{Path where to save images that will be tweet}

\item{post_tweet}{Logical. Whether to really tweet the content}

\item{top_number}{Numeric. Number of best retweeted tweets to show on the graph}

\item{hashtag}{Character. Which Twitter hashtag is followed}

\item{fill}{Vector of colors to be used in the graphics.
Length 1 for all the same, or length 3 resp. for:
top retweeted, tweets per day, contributors per day}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[rtweet:auth_as]{auth_as()}} for details.}

\item{the_month_name}{Character. Month name that appear on graph.
Default to last month name.}
}
\description{
Create a tweet for top retweets and other stats
}
